src/main/java/org/example/spring_jpa_workshop/data/AppUserDaoImpl.java
package org.example.spring_jpa_workshop.data;

import org.example.spring_jpa_workshop.entity.AppUser;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;
import java.util.Collection;

@Repository
public class AppUserDaoImpl implements AppUserDao{

    @PersistenceContext
    EntityManager entityManager;


    @Override
    @Transactional(readOnly = true)
    public AppUser findById(Integer integer) {
        return entityManager.find(AppUser.class, integer);
    }

    @Override
    @Transactional(readOnly = true)
    public Collection<AppUser> findAll() {
        return entityManager.createQuery("select a from AppUser a", AppUser.class).getResultList();
    }

    @Override
    @Transactional
    public AppUser create(AppUser appUser) {
        entityManager.persist(appUser);
        return appUser;
    }

    @Override
    @Transactional
    public AppUser update(AppUser appUser) {
        return entityManager.merge(appUser);
    }

    @Override
    @Transactional
    public void delete(Integer integer) {
        AppUser appUser = entityManager.find(AppUser.class, integer);
        if (appUser != null) entityManager.remove(appUser);
    }
}
